justify app name
