+++ core/urls.py
@@ -6,8 +6,8 @@ urlpatterns = [
     path("", views.home, name="home"),
     path("members/", include("django.contrib.auth.urls")),
     path("members/", include("members.urls")),
-    path("schedules/", include('schedules.urls')),
-    # path("spots/", include("spots.urls")),
+    path("trips/", include("trips.urls")),
+    path("schedules/", include("schedules.urls")),
     path("admin/", admin.site.urls),
     path("accounts/", include("allauth.urls")),
 ]
