+++ schedules/models.py
@@ -1,15 +1,21 @@
 from django.db import models
+from trips.models import Trip
 from django.utils import timezone
-from members.models import Members
 
-class Schedules(models.Model):
-    members = models.ForeignKey(Members, on_delete=models.CASCADE)
-    name = models.CharField(max_length=100)
-    start_date = models.DateTimeField(default=timezone.now)
-    end_date = models.DateTimeField(default=timezone.now)
-    transportation = models.CharField(max_length=20)
+
+class Schedule(models.Model):
+    # trip = models.ForeignKey(Trip, on_delete=models.CASCADE)
+    spot_name = models.CharField(max_length=100)
+    start_time = models.TimeField(default=timezone.now)
+    end_time = models.TimeField(default=timezone.now)
+    note = models.TextField(default="")
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    deleted_at = models.DateTimeField(null=True)
+
+    def soft_delete(self):
+        self.deleted_at = timezone.now()
+        self.save()
 
     def __str__(self):
         return self.spot_name
-
-# members/1/schedule/1/spots/index.html
