+++ templates/pages/home.html
@@ -3,10 +3,10 @@
 
   <ul>
     <li class="inline-block p-2 m-1 bg-yellow-300 border-2">
-      <a href="{% url 'schedules:index' %}">schedules</a>
+      <a href="{% url 'trips:index' %}">trips</a>
     </li>
     <li class="inline-block p-2 m-1 bg-indigo-300 border-2">
-      <a href="#">spots</a>
+      <a href="{% url 'schedules:index' %}">schedules</a>
     </li>
   </ul>
 {% endblock %}
