+++ templates/schedules/new.html
@@ -1,52 +1,87 @@
-{% extends "layouts/base.html" %}
-
-{% block content %}
+{% extends "layouts/base.html" %} {% block content %}
 <div class="h-screen bg-blue-50">
-    <header class="flex justify-start bg-sky-800">
-        <a class="px-2 py-2 text-white bg-sky-800 hover:bg-sky-950" href="{% url 'schedules:index' %}"><i class="fa-solid fa-house-chimney"></i></a>
-        <h1 class="px-2 py-2 text-white bg-sky-800">新增行程</h1>
-    </header>
-    <form action="{% url 'schedules:add' %}" method="post" class="mt-8">
-        {% csrf_token %}
-    
-        <div class="flex flex-col justify-center max-w-md min-h-full px-6 py-5 mx-auto lg:px-10 justify-self-start">
-            <div class="mb-6">
-              <div class="w-full pic">
-                <label for="pic" class="block text-sm font-bold leading-6 text-cyan-800">封面照片</label>
-                <input type="image" id="image" src="https://picsum.photos/370/200">
-              </div>
-            </div>   
-            <div class="mb-6">
-                <label for="name" class="block text-sm font-bold leading-6 text-cyan-800">行程名稱：</label>
-                <input type="text" name="name" id="name" required="required" class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50">
-            </div>
-        
-            <div class="mb-6">
-                <label for="start_date" class="block text-sm font-bold leading-6 text-cyan-800"> 出發日：</label>
-                <input type="date" name="start_date" id="start_date" value="" required="required" class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50">
-            </div>
+	<header class="flex justify-start bg-sky-800">
+		<a
+			class="px-2 py-2 text-white bg-sky-800 hover:bg-sky-950"
+			href=""
+			><i class="fa-solid fa-house-chimney"></i
+		></a>
+		<h1 class="px-2 py-2 text-white bg-sky-800">新增景點</h1>
+	</header>
+	<form action="add" method="post" class="mt-8">
+		{% csrf_token %}
 
-            <div class="mb-6">
-                <label for="end_date" class="block text-sm font-bold leading-6 text-cyan-800">結束日：</label>
-                <input type="date" name="end_date" id="end_date" value="" required="required" class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50">
-            </div>
+		<div
+			class="flex flex-col justify-center max-w-md min-h-full px-6 py-5 mx-auto lg:px-10 justify-self-start"
+		>
+			<div class="mb-6">
+				<label
+					for="spot_name"
+					class="block text-base font-bold leading-6 text-cyan-800"
+					>景點名稱：</label
+				>
+				<input
+					type="text"
+					name="spot_name"
+					id="spot_name"
+					required="required"
+					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
+				/>
+			</div>
 
-            <div class="mb-6">
-                <label for="transportation" class="block text-sm font-bold leading-6 text-cyan-800">交通方式：</label>
-                <select class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50" name="transportation">
-                  <option value="">走路</option>
-                  <option value="">汽車</option>
-                  <option value="">機車</option>
-                  <option value="">大眾運輸</option>
-                </select>
-            </div>
-        </div>
-        <div class="flex justify-center">
-            <button class="px-4 py-2 font-bold text-white rounded bg-sky-800 hover:bg-sky-950">
-                <input type="submit" value="建立新行程">
-            </button>
-        </div>
-    </form>
+			<div class="mb-6">
+				<label
+					for="start_time"
+					class="block text-base font-bold leading-6 text-cyan-800"
+				>
+					抵達時間：</label
+				>
+				<input
+					type="time"
+					name="start_time"
+					id="start_time"
+					value=""
+					required="required"
+					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
+				/>
+			</div>
+
+			<div class="mb-6">
+				<label
+					for="end_time"
+					class="block text-base font-bold leading-6 text-cyan-800"
+					>離開時間：</label
+				>
+				<input
+					type="time"
+					name="end_time"
+					id="end_time"
+					value=""
+					required="required"
+					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
+				/>
+			</div>
+
+			<div class="mb-6">
+				<label
+					for="note"
+					class="block text-base font-bold leading-6 text-cyan-800"
+					>行程筆記：</label
+				>
+				<textarea
+					name="note"
+					id="note"
+					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-800 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
+				></textarea>
+			</div>
+		</div>
+		<div class="flex justify-center">
+			<button
+				class="px-4 py-2 font-bold text-white rounded bg-sky-800 hover:bg-sky-950"
+			>
+				<input type="submit" value="儲存" />
+			</button>
+		</div>
+	</form>
 </div>
 {% endblock %}
-
