--- templates/spots/new.html
@@ -1,87 +1,0 @@
-{% extends "layouts/base.html" %} {% block content %}
-<div class="h-screen bg-blue-50">
-	<header class="flex justify-start bg-sky-800">
-		<a
-			class="px-2 py-2 text-white bg-sky-800 hover:bg-sky-950"
-			href=""
-			><i class="fa-solid fa-house-chimney"></i
-		></a>
-		<h1 class="px-2 py-2 text-white bg-sky-800">新增景點</h1>
-	</header>
-	<form action="add" method="post" class="mt-8">
-		{% csrf_token %}
-
-		<div
-			class="flex flex-col justify-center max-w-md min-h-full px-6 py-5 mx-auto lg:px-10 justify-self-start"
-		>
-			<div class="mb-6">
-				<label
-					for="spot_name"
-					class="block text-base font-bold leading-6 text-cyan-800"
-					>景點名稱：</label
-				>
-				<input
-					type="text"
-					name="spot_name"
-					id="spot_name"
-					required="required"
-					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
-				/>
-			</div>
-
-			<div class="mb-6">
-				<label
-					for="start_time"
-					class="block text-base font-bold leading-6 text-cyan-800"
-				>
-					抵達時間：</label
-				>
-				<input
-					type="time"
-					name="start_time"
-					id="start_time"
-					value=""
-					required="required"
-					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
-				/>
-			</div>
-
-			<div class="mb-6">
-				<label
-					for="end_time"
-					class="block text-base font-bold leading-6 text-cyan-800"
-					>離開時間：</label
-				>
-				<input
-					type="time"
-					name="end_time"
-					id="end_time"
-					value=""
-					required="required"
-					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-300 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
-				/>
-			</div>
-
-			<div class="mb-6">
-				<label
-					for="note"
-					class="block text-base font-bold leading-6 text-cyan-800"
-					>行程筆記：</label
-				>
-				<textarea
-					name="note"
-					id="note"
-					class="block w-full h-12 pl-1 mt-1 mb-3 border-gray-800 rounded-md shadow-sm form-input focus:border-indigo-600 focus:ring focus:ring-cyan-800 focus:ring-opacity-50"
-				></textarea>
-			</div>
-		</div>
-		<div class="flex justify-center">
-			<button
-				class="px-4 py-2 font-bold text-white rounded bg-sky-800 hover:bg-sky-950"
-			>
-				<input type="submit" value="儲存" />
-			</button>
-		</div>
-	</form>
-</div>
-{% endblock %}
