+++ trips/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class SchedulesConfig(AppConfig):
+class TripsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "schedules"
+    name = "trips"
